requirements.txt
anyio==4.3.0
appdirs==1.4.4
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
cssselect==1.2.0
fake-useragent==1.5.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
importlib_metadata==7.1.0
lxml==5.1.0
parse==1.20.1
pyee==11.1.0
pyppeteer==2.0.0
pyquery==2.0.0
python-telegram-bot==21.0.1
requests==2.31.0
requests-html==0.10.0
sniffio==1.3.1
soupsieve==2.5
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==1.26.18
w3lib==2.1.2
websockets==10.4
zipp==3.18.1
